ASoC: madera: Fix madera_dai_set_sysclk() aliasing
In madera_dai_set_sysclk() we are only supposed to be adding or
removing routes to ASYNCCLK, depending whether the selected source
clock is on ASYNCCLK or SYSCLK domain. The routes to SYSCLK are
always required.

The old implementation didn't deal with the fact that all three
SYSCLKs should be treated the same, and likewise all the ASYNCCLKs
are the same. This meant it could be fooled into removing the routes
to SYSCLK by switching clk_id from one SYSCLK to another SYSCLK.

The patch fixes the code to properly collapse the clock choice down
to a selection between SYSCLK and ASYNCCLK domains, and ensure that
only the ASYNCCLK routes are changed.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Signed-off-by: Paul Handrigan <Paul.Handrigan@cirrus.com>

Mot-CRs-fixed: (CR): Beckham: AoU bringup on Moon+Womack
Change-Id: I14fb6d871bf36e408a70a4fbcc18aed487f413f3
Reviewed-on: https://gerrit.mot.com/1139097
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Submit-Approved: Shih-Pei Chen <chensp2@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>